==> Uber October-2023 Basket Analysis/3-Scripts/Notebook 1 - Uber data - Data cleaning and consistency checks - 1.ipynb <==
This notebook contains the initial steps taken for data quality check and consistency of the 'Uber dataset'.

1. Checking shape and data types

2. Looking for missing values

3. Looking for missing data

4. Dealing with negative values

5. Changing data types in some columns

6. Locating outliers

# Importing Analitycal and Visual libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

Creating a path to import and work on data set

path = r'C:\Users\Paola\Desktop\Shaul\Data\CF\Data Immersion\Achievement 6\Uber October-2023 Basket Analysis'

path

Importing Uber data set

df = pd.read_csv(os.path.join(path, '2-Data', 'Original Data', 'uber_data_wrangled.csv'), index_col = False)

Checking shape and columns

df.shape

df.columns

df.info()

Addressing data quality, consistency and initial preparation

# Viewing the missing values
df.isnull().sum()

No missing data.

# Finding duplicates
df_dups = df[df.duplicated()]

df_dups.shape

No duplicates in this data set.

# Checking df for mixing data
for col in df.columns.tolist():
  weird = (df[[col]].applymap(type) != df[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df[weird]) > 0:
    print (col)

No mixed data in this data set.

# This command has been repeated a few times, to check that replacement with mean value was succesful.
df.describe()

There seems to be issues with some inconsistency in the entries. There are many outliers as well as negative amounts in the 'fare_amount' column. Also, some of the columns have numeric values, but they represent options, not values.

Dealing with the negative values.

# Checking the negative values. This and the following two commands have been repeated for the following columns:
# Columns 'fare_amount', 'extra', 'mat_tax', 'tip_amount', 'improvement_surcharge' and 'total_amount'.
df.loc[df['total_amount'] < 0]

The 'fare_amount' column had negative values in 70 rows. This does not make sense, as independently of the length of the trip, any form of payment would have meant a positive value. It is reasonable to ascribe this error to a typo mistake. As it is only 0.0007% of the data, I will use the mean value for imputation to solve this issue.

The 'extra' column had negative values in 19 rows. This does not make sense, as independently of the length of the trip, any form of extras would have meant a positive value. It is reasonable to ascribe this error to a typo mistake. As it is only 0.00019% of the data, I will use the mean value for imputation to solve this issue.

The 'mta_tax' column had negative values in 66 rows. This does not make sense, as independently of the length of the trip, any form of tax would have meant a positive value. It is reasonable to ascribe this error to a typo mistake. As it is only 0.00066% of the data, I will use the mean value for imputation to solve this issue.

The 'tip_amount' column had negative values in 4 rows. This does not make sense, as independently of the length of the trip, any form of tip would have meant a positive value. It is reasonable to ascribe this error to a typo mistake. As it is only 0.0001% of the data, I will use the mean value for imputation to solve this issue.

The 'improvement_surcharge' column had negative values in 70 rows. This does not make sense, as independently of the length of the trip, any form of payment would have meant a positive value. It is reasonable to ascribe this error to a typo mistake. As it is only 0.0007% of the data, I will use the mean value for imputation to solve this issue.

The 'improvement_surcharge' column had negative values in 70 rows. This does not make sense, as independently of the length of the trip, any form of payment would have meant a positive value. It is reasonable to ascribe this error to a typo mistake. As it is only 0.0007% of the data, I will use the mean value for imputation to solve this issue.

# Replacing  negative outliers with the mean. 
df.loc[df['total_amount'] < 0, 'total_amount'] = np.nan

# Negative outliers solved => same operation repeated for all columns just named above.
df['total_amount'].fillna((df['total_amount'].mean()), inplace=True)

Changing data types from numeric to string.

# Changing the following columns data types to string (text): VendorID, passenger_count, payment_type, RateCodeID. 
# Even though written in numbers, they represent text options, not numbers. This comand has been repeated for all the abovementioned columns.
df['payment_type'] = df['payment_type'].astype('str')

The numeric columns are: trip_distance, fare_amount, extra, mta_tax, tip_amount, tolls_amount, mprovement_surcharge, total_amount.
Pickup_time and dropoff_time alre also numeric, but currently their data is shown as object.

# Checking the outliers in the max values.
# This command has been repeated with different values, and with both columns 'trip distance' and 'fare amount', to get the report below.
df.loc[df['fare_amount'] > 300]

From 100000 trips, there are:
Just 2 that exceed 100 miles (entries 75653 and 99665)
Another 2 that exceed 50 miles (entries 35136 and 94698)
Another 551 entries that exceed 20 miles
Another 2355 entries that exceed 15 miles. In total, 2910 entries.

There are 2921 trips that exceed $50 dollars. From those, 64 that exceed $100 dollars. From the latter, 26 that exceed $150 dollars. Lastly, from those 10 exceed $200 dollars. Only four trips exceed $300 dollars.

There seems to be no relation between length of trip and fare amount, as I can observe some relatively short trips costing very expensive, and some long trips costing less. Also, observing pick up and drop times it seems that fare amount is judged by time in the cab, not length of trip. I would go to my superior to check this.

# Exporting 'df' to csv
df.to_csv(os.path.join(path, '2-Data','Prepared Data', 'uber_checked.csv'))
