==> Uber October-2023 Basket Analysis/3-Scripts/Notebook 2 - Uber data - Data cleaning and consistency checks - 2.ipynb <==
# Importing Analitycal and Visual libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

Creating a path to import and work on data set

path = r'C:\Users\Paola\Desktop\Shaul\Data\CF\Data Immersion\Achievement 6\Uber October-2023 Basket Analysis'

path

Importing Uber data set

df = pd.read_csv(os.path.join(path, '2-Data', 'Prepared Data', 'uber_checked.csv'), index_col = False)

Checking shape and columns

df.shape

df.columns

# For some reason, I had an unnamed column when I imported this dataset.
# I used this command to drop it, and then rerun the last two commands to check the results.
df = df.loc[:, ~df.columns.str.contains('^Unnamed')]

df.describe()

# Following my comments in the first notebook: there are 10 values in the 'fare amount' column that greatly skew the data.
# As they do not pose a significance from an analytical perspective, I am going to remove those from the dataset.
df.loc[df['fare_amount'] > 200]

# Dropping the 10 outliers
df = df[df.fare_amount < 201]

df.shape

df.describe()

# Exporting 'df' to csv
df.to_csv(os.path.join(path, '2-Data','Prepared Data', 'uber_checked2.csv'), index=False)
