==> Uber October-2023 Basket Analysis/3-Scripts/Notebook 4 - Uber data - dealing with datetime variables.ipynb <==
This notebook contains:
    
1) commands to limit trips to $50 as not to skew the data analysis

2) creating the new variables, not saved in the previous notebook (grouping trips by length and fare amount)

3) converting date and time variables from 'object' to date and time values

4) exporting updated notebook to file


# Importing Analitycal and Visual libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
from datetime import datetime 

Creating a path

path = r'C:\Users\Paola\Desktop\Shaul\Data\CF\Data Immersion\Achievement 6\Uber October-2023 Basket Analysis'

path

df = pd.read_csv(os.path.join(path, '2-Data', 'Prepared Data', 'uber_checked2.csv'), index_col = False)

# Checking the dataset
df.shape

df.columns

# There are 2921 trips that exceed $50 dollars. From those, 64 that exceed $100 dollars. 
# From the latter, 26 that exceed $150 dollars. Lastly, from those 10 exceed $200 dollars. 
# Only four trips exceed $300 dollars. I am limiting the visual analysis to trip costing up to $50, to avoid skewing
# the data unnecessarily.
df = df.loc[df['fare_amount'] < 51]

df.shape

# NOTE: I am repeating these steps to create these variables, as I did not save them in the previous notebook.
# Creating a new variable to categorise trips by length
df.loc[df['trip_distance'] < 5, 'Trip category'] = 'Short trip'

df.loc[(df['trip_distance'] >= 5) & (df['trip_distance'] <10), 'Trip category'] = 'Normal trip'

df.loc[df['trip_distance'] >= 10, 'Trip category'] = 'Longer trip'

# Checking value counts for the new variable
df['Trip category'].value_counts(dropna = False)

# NOTE: I am repeating these steps to create this variable for the same reason as stated above.
# Creating a new variable to categorise trips by length
df.loc[df['fare_amount'] < 10, 'Fare category'] = 'Cheap trip'

df.loc[(df['fare_amount'] >= 10) & (df['fare_amount'] <20), 'Fare category'] = 'Midprice trip'

df.loc[df['fare_amount'] >= 20, 'Fare category'] = 'Expensive trip'

# Checking value counts for the new variable
df['Fare category'].value_counts(dropna = False)

Dealing with the date and time variables

df.pickup_date.describe()

# Renaming/Standardising column name: from 'dropff_date' to 'dropoff_date'
df.rename(columns = {'dropff_date' : 'dropoff_date'}, inplace = True)

# I have repeated this command to create new four variables.
#These are: 'pickup date new', 'dropoff date new', 'pickup time new', 'dropoff time new'.
df['dropoff_date_new'] = pd.to_datetime(df['dropoff_date'], format = '%d/%m/%Y %H:%M').dt.date

df.pickup_date_new.describe(datetime_is_numeric=True)

df.dropoff_date_new.describe(datetime_is_numeric=True)

df.pickup_time_new.describe(datetime_is_numeric=True)

df.dropoff_time_new.describe(datetime_is_numeric=True)

df.columns

# Exporting 'df' to csv
df.to_csv(os.path.join(path, '2-Data','Prepared Data', 'uber_limited_to_usd50_trips.csv'), index=False)
